Refs #115. Moved entry point code to __main__ module
